# acs/web/openapi.yaml
openapi: "3.0.3"
info:
  title: Agent Cron Scheduler API
  description: |
    REST API for the Agent Cron Scheduler (ACS) daemon.

    ACS is a multi-step workflow scheduler with a built-in HTTP API and web UI.
    Workflows can be identified by either their UUID or their unique name in all
    endpoints that accept an `{id}` path parameter.
  version: "4.0.0"
  license:
    name: MIT
...
paths:
...
  /api/restart:
    post:
      operationId: restart
      summary: Restart the daemon
      description: |
        Spawns a new daemon process and then shuts down the current one after a
        short delay (500ms) to allow the response to be sent.
      tags: [System]
      responses:
        "200":
          description: Restart initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: Restart initiated
        "500":
          description: Failed to spawn new process or determine executable path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

# ===========================================================================
# Components
# ===========================================================================
components:
...
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Shutdown initiated
    ErrorResponse:
      type: object
      required: [error, message]
      description: Standard error envelope returned by all error responses.
      properties:
        error:
          type: string
          description: Machine-readable error code.
          example: not_found
        message:
          type: string
          description: Human-readable error description.
          example: "Workflow with id '550e8400-e29b-41d4-a716-446655440000' not found"
...